moved state to home until we can set up redux
